test: move Playwright browser binary tests
Move the product-shipped Playwright browser binary resolver and layout validation tests out of test_common_utils and into a renderer-owned test module.

This matches the refactor that localized the Playwright browser binary helpers in pdf_renderer.py, so common utility tests no longer import browser-PDF-specific renderer APIs.

Verification: py_compile on the affected test and renderer files passed; git diff --check passed. uv-backed gates remain unavailable because uv is not on PATH in this environment.
